+++ tokio-signal/src/lib.rs
@@ -33,7 +33,7 @@
 //! async fn main() -> Result<(), Box<dyn std::error::Error>> {
 //!     // Create an infinite stream of "Ctrl+C" notifications. Each item received
 //!     // on this stream may represent multiple ctrl-c signals.
-//!     let ctrl_c = tokio_signal::CtrlC::new().await?;
+//!     let ctrl_c = tokio_signal::CtrlC::new()?;
 //!
 //!     // Process each ctrl-c as it comes in
 //!     let prog = ctrl_c.for_each(|_| {
@@ -60,7 +60,7 @@
 //! async fn main() -> Result<(), Box<dyn std::error::Error>> {
 //!     // Create an infinite stream of "Ctrl+C" notifications. Each item received
 //!     // on this stream may represent multiple ctrl-c signals.
-//!     let ctrl_c = tokio_signal::CtrlC::new().await?;
+//!     let ctrl_c = tokio_signal::CtrlC::new()?;
 //!
 //!     // Process each ctrl-c as it comes in
 //!     let prog = ctrl_c.for_each(|_| {
@@ -72,7 +72,7 @@
 //!
 //!     // Like the previous example, this is an infinite stream of signals
 //!     // being received, and signals may be coalesced while pending.
-//!     let stream = Signal::new(SIGHUP).await?;
+//!     let stream = Signal::new(SIGHUP)?;
 //!
 //!     // Convert out stream into a future and block the program
 //!     let (signal, _signal) = stream.into_future().await;
@@ -83,10 +83,6 @@
 
 #[macro_use]
 extern crate lazy_static;
-
-use futures_core::future::Future;
-use std::io;
-use std::pin::Pin;
 
 mod ctrl_c;
 mod registry;
@@ -100,8 +96,5 @@ mod os {
 
 pub mod unix;
 pub mod windows;
-
-/// A future whose output is `io::Result<T>`
-pub type IoFuture<T> = Pin<Box<dyn Future<Output = io::Result<T>> + Send>>;
 
 pub use ctrl_c::CtrlC;